# server/openapi.yaml
openapi: 3.1.0
info:
  title: Aria API
  version: 2.0.0
  description: |
    Aria v2 music-server API. Paths and JSON shapes are byte-compatible with the
    legacy Node server; v2 adds GET /healthz, GET /api/events (SSE), and optional
    limit/offset pagination on GET /api/tracks.

    ID scheme: track `id` = sha1 hex of the file path relative to MUSIC_DIR;
    `albumId` = sha1 hex of lowercase(album directory + "\0" + album title), the
    directory being the file's folder with a CD1/Disc 2 subfolder folded into its
    parent, and the album artist replacing the (empty) directory at library root.

    Error convention: validation failures are 400 with `{"error": "<reason>"}`;
    not-found is a bare 404 (text/plain body "Not Found", no JSON);
    upstream MusicBrainz failures are 502 with `{"error": "..."}`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/tracks/{id}:
    patch:
      tags: [edits]
      summary: Edit track metadata (DB override)
      description: |
        Overrides beat file tags and enrichment. Setting a field to `null` clears
        that single override. Returns the surviving override map for the track —
        NOT the updated track object; may be `{}` if the patch cleared everything.
      operationId: patchTrack
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrackEdits' }
      responses:
        '200':
          description: Current override map for this track.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackEdits' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/identify/artist/{name}:
    get:
      tags: [enrichment]
      summary: MusicBrainz artist candidates
      operationId: identifyArtist
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Up to 8 candidates.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ArtistCandidate' }
        '502': { $ref: '#/components/responses/Upstream' }
  /api/artist/{name}/reidentify:
    post:
      tags: [enrichment]
      summary: Re-enrich an artist, optionally pinned to a chosen MBID
      description: Drops the cached entry and re-enriches. Returns the fresh Artist entry, or `{}` when re-enrichment produced nothing.
      operationId: reidentifyArtist
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mbid:
                  oneOf:
                    - { $ref: '#/components/schemas/Mbid' }
                    - { type: 'null' }
      responses:
        '200':
          description: Fresh artist entry, or empty object.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/Artist' }
                  - { type: object, additionalProperties: false }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { $ref: '#/components/responses/Upstream' }
  /api/identify/album/{albumId}:
    get:
      tags: [enrichment]
      summary: MusicBrainz release candidates for an album
      operationId: identifyAlbum
      parameters:
        - { name: albumId, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      responses:
        '200':
          description: Up to 8 candidates.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlbumCandidate' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/Upstream' }
  /api/album/{albumId}/reidentify:
    post:
      tags: [enrichment]
      summary: Re-enrich an album, optionally pinned to a chosen MBID
      description: |
        Clears cached per-track credits and album info, re-pulls from MusicBrainz.
        Existing cover-art file is kept. Returns the repopulated AlbumInfo, or `{}`
        when nothing was found.
      operationId: reidentifyAlbum
      parameters:
        - { name: albumId, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mbid:
                  oneOf:
                    - { $ref: '#/components/schemas/Mbid' }
                    - { type: 'null' }
      responses:
        '200':
          description: Fresh album info, or empty object.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/AlbumInfo' }
                  - { type: object, additionalProperties: false }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/Upstream' }

  # ---- edits -----------------------------------------------------------------
  /api/albums/{albumId}:
    patch:
      tags: [edits]
      summary: Edit album metadata (DB override)
      description: |
        `album`, `albumArtist`, `genre`, `year` fan out onto the album's tracks in
        /api/tracks; `releaseType`, `label`, `date`, `country`, `blurb` stay
        album-level (served via /info and /edits). `null` clears one field.
        Returns the surviving override map, not the album.
      operationId: patchAlbum
      parameters:
        - { name: albumId, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlbumEdits' }
      responses:
        '200':
          description: Current override map for this album.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumEdits' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/artists/{name}:
    patch:
      tags: [edits]
      summary: Edit artist/person metadata (DB override)
      description: Any name is allowed (max 200 chars), including names not in the library — edits make MB-unknown names browsable. Returns the surviving override map.
      operationId: patchArtist
      parameters:
        - { name: name, in: path, required: true, schema: { type: string, maxLength: 200 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArtistEdits' }
      responses:
        '200':
          description: Current override map for this artist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistEdits' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/edits/{kind}/{key}:
    get:
      tags: [edits]
      summary: Editor support — original values and current overrides
      description: |
        `original` comes from file tags + enrichment only, never from edits; every
        editable field is present, null when unset. `overrides` is the raw stored
        override map (may be `{}`). Field sets per kind:
        track: title, artist, album, albumArtist, genre, year, trackNo, discNo,
        composer, work, movement, conductor, orchestra.
        album: album, albumArtist, genre, year, releaseType, label, date, country, blurb,
        state, disambiguation, locked.
        artist: type, area, born, died, image, bio (cache-only, never hits the network).
        For kind=artist the key may be the Latin display name; it is resolved back to
        the stored (original-script) name before anything is read or written.
      operationId: getEdits
      parameters:
        - name: kind
          in: path
          required: true
          schema: { type: string, enum: [track, album, artist] }
        - name: key
          in: path
          required: true
          schema: { type: string }
          description: Track id, albumId, or artist name.
      responses:
        '200':
          description: Original values + overrides.
          content:
            application/json:
              schema:
                type: object
                required: [original, overrides]
                properties:
                  original: { type: object }
                  overrides: { type: object }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- tags ------------------------------------------------------------------
...
components:
  responses:
...
    BadRequest:
      description: Validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found. Plain-text body ("Not Found"), no JSON.
      content:
        text/plain:
          schema: { type: string }
    Upstream:
      description: MusicBrainz unavailable or upstream lookup failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Sha1:
      type: string
      pattern: '^[0-9a-f]{40}$'
    Mbid:
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Artist:
      type: object
      description: Enriched artist/person entry (all fields optional — whatever the sources yielded), with DB edits overlaid.
      properties:
        type: { type: [string, 'null'], description: 'MusicBrainz artist type, e.g. "Person", "Group".' }
        area: { type: [string, 'null'] }
        born: { type: [string, 'null'], description: Life-span begin date (YYYY or YYYY-MM-DD). }
        died: { type: [string, 'null'] }
        bio: { type: string, description: Wikipedia summary extract. }
        url: { type: [string, 'null'], description: Wikipedia page URL. }
        image: { type: string, description: 'Portrait URL (Wikipedia preferred, Deezer fallback, edits win).' }
        imgSrc: { type: string, enum: [wikipedia, deezer] }
        imgCheckedAt: { type: string, format: date-time, description: Set once the face-upgrade pass has tried this entry. }
        members:
          type: array
          items: { type: string }
          description: When this artist is a group — member names.
        bands:
          type: array
          items: { type: string }
          description: When this artist is a person — bands they belong(ed) to.
        similar:
          type: array
          maxItems: 12
          items:
            type: object
            required: [name, image]
            properties:
              name: { type: string }
              image: { type: [string, 'null'] }
        discography:
          type: array
          items: { $ref: '#/components/schemas/DiscographyItem' }
        discographyAt: { type: string, format: date-time }
        mbid: { type: string, description: MusicBrainz artist MBID, when one was resolved. }
        nameLatin:
          type: string
          description: >
            Latin-script display spelling of the entry's key. Absent when the
            stored name is already Latin, or when no source yielded one — the
            original script is never overwritten.
        nameLatinSource:
          type: string
          enum: [alias-en, alias, sortname, translit, mbname, same, none]
          description: How `nameLatin` was resolved; `translit` is approximate.
    DiscographyItem:
      type: object
      required: [title, cover, date, type]
      properties:
        title: { type: string }
        cover: { type: [string, 'null'] }
        date: { type: [string, 'null'], description: YYYY-MM-DD. }
        type: { type: string, description: 'Lowercase Deezer record type (album, single, ep, compilation).' }
...
    AlbumInfo:
      type: object
      description: On-demand album depth; only found fields are present. Edits overlay label/date/country/blurb.
      properties:
        label: { type: [string, 'null'] }
        date: { type: [string, 'null'], description: MusicBrainz release date. }
        country: { type: [string, 'null'] }
        mbType: { type: [string, 'null'], description: MB release-group primary type. }
        mbSecondary:
          type: array
          items: { type: string }
          description: MB release-group secondary types (Compilation, Live, ...).
        blurb: { type: string, description: Wikipedia summary of the album. }
        url: { type: [string, 'null'], description: Wikipedia page URL. }
    ArtistCandidate:
      type: object
      required: [mbid, name, type, area, disambiguation, score]
      properties:
        mbid: { $ref: '#/components/schemas/Mbid' }
        name: { type: string }
        type: { type: [string, 'null'] }
        area: { type: [string, 'null'] }
        disambiguation: { type: [string, 'null'] }
        score: { type: [integer, 'null'] }
    AlbumCandidate:
      type: object
      required: [mbid, title, artist, date, country, tracks, score]
      properties:
        mbid: { $ref: '#/components/schemas/Mbid' }
        title: { type: string }
        artist: { type: [string, 'null'], description: Joined artist-credit string. }
        date: { type: [string, 'null'] }
        country: { type: [string, 'null'] }
        tracks: { type: [integer, 'null'], description: MB track count. }
        score: { type: [integer, 'null'] }
    TrackEdits:
      type: object
      description: Track override fields. In a PATCH request, null clears a field; other values must be non-empty trimmed strings (max 300) or ints 0-3000 for year/trackNo/discNo. In responses, cleared fields are absent.
      additionalProperties: false
      properties:
        title: { type: [string, 'null'] }
        artist: { type: [string, 'null'] }
        album: { type: [string, 'null'] }
        albumArtist: { type: [string, 'null'] }
        genre: { type: [string, 'null'] }
        year: { type: [integer, 'null'], minimum: 0, maximum: 3000 }
        trackNo: { type: [integer, 'null'], minimum: 0, maximum: 3000 }
        discNo: { type: [integer, 'null'], minimum: 0, maximum: 3000 }
        composer: { type: [string, 'null'] }
        work: { type: [string, 'null'] }
        movement: { type: [string, 'null'] }
        conductor: { type: [string, 'null'] }
        orchestra: { type: [string, 'null'] }
    AlbumEdits:
      type: object
      description: Album override fields; same null-clears semantics as TrackEdits. blurb allows up to 4000 chars.
      additionalProperties: false
      properties:
        album: { type: [string, 'null'] }
        albumArtist: { type: [string, 'null'] }
        genre: { type: [string, 'null'] }
        year: { type: [integer, 'null'], minimum: 0, maximum: 3000 }
        releaseType:
          oneOf:
            - { type: string, enum: [Album, EP, Single, Live, Compilation] }
            - { type: 'null' }
        label: { type: [string, 'null'] }
        date: { type: [string, 'null'] }
        country: { type: [string, 'null'] }
        blurb: { type: [string, 'null'], maxLength: 4000 }
        artSource:
          oneOf:
            - { type: string, enum: [file, api, custom] }
            - { type: 'null' }
        state:
          description: '`local` means the user''s metadata is authoritative and no MusicBrainz entity is pinned.'
          oneOf:
            - { type: string, enum: [matched, local] }
            - { type: 'null' }
        disambiguation: { type: [string, 'null'] }
        locked: { type: [boolean, 'null'] }
    ArtistEdits:
      type: object
      description: Artist override fields; same null-clears semantics. bio allows up to 4000 chars.
      additionalProperties: false
      properties:
        type: { type: [string, 'null'] }
        area: { type: [string, 'null'] }
        born: { type: [string, 'null'] }
        died: { type: [string, 'null'] }
        image: { type: [string, 'null'] }
        bio: { type: [string, 'null'], maxLength: 4000 }
...